refactor: quiz 컴포넌트 구조 정리 (디렉토리 분리·불필요 컴포넌트 인라인) (#48)
