--- src/components/quiz/ExamSubmitModal/index.ts
@@ -1,1 +1,0 @@
-export { ExamSubmitModal } from './ExamSubmitModal'
