+++ src/components/quiz/exam/CheatingWarningModal/index.ts
