+++ src/components/quiz/exam/questions/MultipleChoiceQuestion/index.ts
