+++ src/components/quiz/index.ts
@@ -1,6 +1,6 @@
-export { ExamThumbnail } from './ExamThumbnail'
-export { QuizCard } from './QuizCard'
-export { QuizCardSkeleton } from './QuizCardSkeleton'
-export { QuizList } from './QuizList'
-export { QuizListErrorBoundary } from './QuizListErrorBoundary'
-export { QuizCodeModal } from './QuizCodeModal'
+export { ExamThumbnail } from './list/ExamThumbnail'
+export { QuizCard } from './list/QuizCard'
+export { QuizCardSkeleton } from './list/QuizCardSkeleton'
+export { QuizList } from './list/QuizList'
+export { QuizListErrorBoundary } from './list/QuizListErrorBoundary'
+export { QuizCodeModal } from './list/QuizCodeModal'
