+++ src/components/quiz/list/ExamThumbnail/ExamThumbnail.tsx
