+++ src/components/quiz/list/ExamThumbnail/index.ts
