+++ src/components/quiz/list/QuizCardSkeleton/QuizCardSkeleton.tsx
