+++ src/components/quiz/list/QuizCodeModal/QuizCodeModal.tsx
@@ -3,7 +3,7 @@ import { useNavigate } from 'react-router'
 import { Modal } from '@/components/common/Modal'
 import { Input } from '@/components/common/Input'
 import { Button } from '@/components/common/Button'
-import { ExamThumbnail } from '@/components/quiz/ExamThumbnail'
+import { ExamThumbnail } from '@/components/quiz/list/ExamThumbnail'
 import { ROUTES } from '@/constants/routes'
 import { HTTP_STATUS } from '@/constants/httpStatus'
 import { useCheckCode } from '@/features/exams/deployment-check-code'
